+++ ctp2_code/gs/gameobj/bldque.cpp
@@ -623,10 +623,10 @@ bool BuildQueue::BuildFront(sint32 &shieldstore, CityData *cd, const MapPoint &p
 				double penalty =
 					static_cast<double>(g_theConstDB->Get(0)->GetChangeCurrentlyBuildingItemPenalty());
 
-				penalty = std::min<float>
+				penalty = std::min<double>
 								  (
 								   1.0,
-								   std::max<float>(0.0, 1.0 - penalty * 0.01)
+								   std::max<double>(0.0, 1.0 - penalty * 0.01)
 								  );
 
 				s = static_cast<sint32>(static_cast<double>(shieldstore) * penalty);
